+++ fairmlhealth/performance_metrics.py
@@ -1,8 +1,5 @@
 """
 """
-import copy
-from functools import partial
-import logging
 from numbers import Number
 import numpy as np
 import pandas as pd