+++ test/test_fairnessmetrics.py
@@ -2,11 +2,10 @@
 Validation tests for fairmlhealth
 """
 
-import fairmlhealth.fairness_metrics as fm
+import fairmlhealth.__fairness_metrics as fm
 from numpy import isnan
 import pandas as pd
 import pytest
-from fairmlhealth import stat_utils
 from .__utils import synth_dataset